refactor(modules): move ChannelMerger/FourierEmb into layers.py; drop merger.py; rewire dance imports
